refactor: Consolidate evaluation report files and eliminate redundancy
- Create centralized evaluation-report-types.ts with all type definitions
- Consolidate functionality into evaluation-report-db.ts with localStorage fallback
- Remove redundant evaluation-report-storage.ts file
- Remove wrapper evaluation-report-page.tsx file
- Update all components to use consolidated imports from evaluation-report-db.ts
- Update routing to point directly to refactored EvaluationReportPage component
- Maintain full functionality while eliminating code duplication

File changes:
- NEW: src/lib/evaluation-report-types.ts (centralized types)
- ENHANCED: src/lib/evaluation-report-db.ts (added high-level functions with fallback)
- REMOVED: src/lib/evaluation-report-storage.ts (redundant)
- REMOVED: src/pages/evaluation-report-page.tsx (redundant wrapper)
- UPDATED: All component imports to use consolidated module
- UPDATED: App.tsx routing to use refactored component directly

The evaluation report functionality now has a clean, single-source architecture while maintaining all existing features including database storage with localStorage fallback.
