--- src/pages/evaluation-report-page.tsx
@@ -1,2 +1,0 @@
-// Re-export the refactored component
-export { EvaluationReportPage as default } from './evaluation-reports';
